+++ app/Core/Autoloader.php
@@ -12,17 +12,15 @@ class Autoloader
         // Add vendor classes
         if (strpos($className, '_')) {
             $className = 'vendor/' . str_replace("_", "/", $className);
-            require dirname(__DIR__, 2) . DIRECTORY_SEPARATOR . str_replace('\\', DIRECTORY_SEPARATOR, $className) . '.php';
+            require dirname(__DIR__, 2) . DIRECTORY_SEPARATOR . str_replace('\\', DIRECTORY_SEPARATOR,
+                    $className) . '.php';
+        }
+        if (strpos($className, 'Symfony') !== false) {
+            require dirname(__DIR__, 2) . DIRECTORY_SEPARATOR . str_replace('\\', DIRECTORY_SEPARATOR,
+                    'vendor/'. $className) . '.php';
         } else {
             // Add other classes
             require dirname(__DIR__, 1) . DIRECTORY_SEPARATOR . str_replace('\\', DIRECTORY_SEPARATOR, $className) . '.php';
-
-            // Other example method
-//            $file = dirname(__DIR__, 1) . '/'  . $className . '.php';
-//            $file = str_replace('\\', DIRECTORY_SEPARATOR, $file);
-//            if (file_exists($file)) {
-//                include $file;
-//            }
         }
     }
 }