+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-##  (2021-12-08)
+##  (2021-12-12)
 
 [View Release](ssh://git@github.com/mhaagen85/webshop.git/commits/tag/)
 
@@ -46,5 +46,6 @@
 *  [FEATURE] Added test trait *(martijn.vanhaagen@experius.nl)*
 *  [FEATURE] Added return type *(martijn.vanhaagen@experius.nl)*
 *  [FEATURE] Added migrations *(martijn.vanhaagen@experius.nl)*
+*  [FEATURE] Added console command + migrations *(martijn.vanhaagen@experius.nl)*